+++ nw/ARENA/SC/ScriptViewPointState.cpp
@@ -1,0 +1,5 @@
+#define LAST_H__VIEW
+#include "game.h"
+#include "h/vehicle.h"
+
+#include "ScriptViewPointState.inl"