+++ nw/VIEW/TerrainWaterlineState.cpp
@@ -1,0 +1,5 @@
+#ifndef __VIEW_CPP__
+#include "_view.h"
+#endif
+
+#include "TerrainWaterlineState.inl"